requirements.txt
pandas 
requests
beautifulsoup4
faker
python-dotenv
google-genai

